+++ Source/Anvil/Views/Notification/TrayIcon.xaml.cs
@@ -9,26 +9,27 @@ namespace Anvil.Views.Notification
 {
     public partial class TrayIcon : TaskbarIcon, IStartable, IDisposable
     {
-        private readonly TrayIconViewModel mViewModel;
+        private readonly Func<TrayIconViewModel> mViewModelFactory;
 
-        public TrayIcon(TrayIconViewModel viewModel)
+        public TrayIcon(Func<TrayIconViewModel> viewModelFactory)
         {
             InitializeComponent();
 
+            mViewModelFactory = viewModelFactory;
             Icon = Properties.Resources.AnvilIcon;
-            DataContext = mViewModel = viewModel;
         }
 
         public new void Dispose()
         {
             Visibility = Visibility.Collapsed;
-            mViewModel.Dispose();
+            (DataContext as IDisposable)?.Dispose();
 
             base.Dispose();
         }
 
         public void Start()
         {
+            DataContext = mViewModelFactory();
             Visibility = Visibility.Visible;
         }
     }