+++ Source/Anvil/Views/Notification/TrayIconViewModel.cs
@@ -1,8 +1,15 @@
 ﻿using System;
+using System.Collections.ObjectModel;
+using System.Reactive.Linq;
 using System.Windows;
 
 using Anvil.Framework.ComponentModel;
 using Anvil.Framework.MVVM;
+using Anvil.Services;
+using Anvil.Services.Data;
+
+using DynamicData;
+using DynamicData.Binding;
 
 using ReactiveUI;
 
@@ -10,20 +17,33 @@ namespace Anvil.Views.Notification
 {
     public sealed class TrayIconViewModel : DisposableViewModel
     {
-        public TrayIconViewModel()
+        private readonly ReadOnlyObservableCollection<TrayIconLaunchGroupTreeNode> mLaunchGroups;
+
+        public TrayIconViewModel(IDataService dataService, ILaunchService launchService)
         {
             ExitCommand = ReactiveCommand.Create();
             ExitCommand.Subscribe(_Exit).TrackWith(Disposables);
 
             ShowMainWindowCommand = ReactiveCommand.Create();
             ShowMainWindowCommand.Subscribe(_ShowMainWindow).TrackWith(Disposables);
+
+            dataService
+                .LaunchGroups.Connect()
+                .ObserveOnDispatcher()
+                .TransformToTree(grp => grp.ParentGroupId ?? -1)
+                .Transform(grp => new TrayIconLaunchGroupTreeNode(grp, dataService, launchService))
+                .Sort(SortExpressionComparer<TrayIconLaunchGroupTreeNode>.Ascending(x => x.Model.Name))
+                .Bind(out mLaunchGroups).DisposeMany()
+                .Subscribe().TrackWith(Disposables);
         }
 
         public ReactiveCommand<object> ExitCommand { get; }
 
+        public ReadOnlyObservableCollection<TrayIconLaunchGroupTreeNode> LaunchGroups => mLaunchGroups;
+
         public ReactiveCommand<object> ShowMainWindowCommand { get; }
 
-        private void _Exit(object _)
+        private static void _Exit(object _)
         {
             Application.Current.Shutdown();
         }